Validando regras de negocio e customizando exceptions

+++ src/main/java/io/github/robledocastro/libraryapi/service/AutorService.java
@@ -1,8 +1,11 @@
 package io.github.robledocastro.libraryapi.service;
 
 import io.github.robledocastro.libraryapi.controller.dto.AutorDTO;
+import io.github.robledocastro.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.robledocastro.libraryapi.model.Autor;
 import io.github.robledocastro.libraryapi.repository.AutorRepository;
+import io.github.robledocastro.libraryapi.repository.LivroRepository;
+import io.github.robledocastro.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,12 +16,17 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository){
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository){
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -26,6 +34,7 @@ public class AutorService {
         if(autor.getId() == null){
             throw new IllegalArgumentException("Para atualizar, é necessário que o autor já esteja salvo na base.");
         }
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -34,6 +43,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um Autor que possui livros cadastrados!");
+        }
+
         repository.delete(autor);
     }
 
@@ -51,6 +64,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
 
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
